Add support for float seconds in typesAndWaits() method

+++ src/Drivers/Tests/FakeDriver.php
@@ -173,10 +173,10 @@ class FakeDriver implements DriverInterface, VerifiesService
     /**
      * Send a typing indicator and wait for the given amount of seconds.
      * @param IncomingMessage $matchingMessage
-     * @param int $seconds
+     * @param float $seconds
      * @return mixed
      */
-    public function typesAndWaits(IncomingMessage $matchingMessage, int $seconds)
+    public function typesAndWaits(IncomingMessage $matchingMessage, float $seconds)
     {
         $this->types($matchingMessage);
     }
